+++ src/Rxn/Framework/Tests/Http/OpenApi/Fixture/v1/ProductsController.php
@@ -2,7 +2,6 @@
 
 namespace Rxn\Framework\Tests\Http\OpenApi\Fixture\v1;
 
-use Rxn\Framework\Http\Request;
 use Rxn\Framework\Tests\Http\OpenApi\Fixture\BaseController;
 
 final class ProductsController extends BaseController
@@ -10,7 +9,7 @@ final class ProductsController extends BaseController
     /**
      * Show a product by id.
      */
-    public function show_v1(int $id, string $filter = 'all', bool $verbose = false, ?Request $request = null): array
+    public function show_v1(int $id, string $filter = 'all', bool $verbose = false): array
     {
         return [];
     }